untangle circular deps
