+++ failgood/common/src/failgood/TestDependency.kt
@@ -1,0 +1,7 @@
+package failgood
+
+import kotlin.reflect.KProperty
+
+interface TestDependency<T> {
+    operator fun getValue(owner: Any?, property: KProperty<*>): T
+}
