--- failgood/jvm/src/failgood/TestDependency.kt
@@ -1,11 +1,0 @@
-package failgood
-
-import kotlinx.coroutines.Deferred
-import kotlinx.coroutines.runBlocking
-import kotlin.reflect.KProperty
-
-class TestDependency<T>(private val dependency: Deferred<Result<T>>) {
-    operator fun getValue(owner: Any?, property: KProperty<*>): T {
-        return runBlocking { dependency.await().getOrThrow() }
-    }
-}
